% Generated from roxygen comments for offline documentation. Do not edit by hand.
\name{PronMensual}
\alias{PronMensual}
\title{Resumir pronosticos por mes}
\usage{
PronMensual(...)
}
\arguments{
  \item{seleccion}{Objeto `pronostico_racafe` post-seleccion.}
  \item{columna}{Columna. `NULL` combina todas.}
  \item{incluir_pronosticos}{Logico. Incluir columna de pronostico puntual.}
}
\value{`tibble` con resumen mensual.}
\description{
Resumir pronosticos por mes
}
