% Generated from roxygen comments for offline documentation. Do not edit by hand.
\name{PronSerie}
\alias{PronSerie}
\title{Construir tabla detallada de serie pronosticada}
\usage{
PronSerie(...)
}
\arguments{
  \item{seleccion}{Objeto `pronostico_racafe` post-seleccion.}
  \item{columna}{Columna. `NULL` combina todas.}
}
\value{`tibble` con fecha, pronostico e intervalos.}
\description{
Construir tabla detallada de serie pronosticada
}
